make it better

+++ src/app/Master/interested/interested.component.ts
@@ -6,7 +6,7 @@ import { DynamicScriptLoaderService } from '../../services/dynamic-script-loader
 import { AuthService } from "../../services/auth.service";
 declare const $: any;
 declare const M: any;
-declare const swal: any;
+import Swal from 'sweetalert2/dist/sweetalert2.js';
 @Component({
   selector: 'app-interested',
   templateUrl: './interested.component.html',
@@ -55,7 +55,7 @@ export class InterestedComponent implements OnInit {
       this.registerForm.value;
     this.request.addInterests(this.registerForm.value).subscribe((res: any) => {
       if (res.status == 'success') {
-        swal("Added Sucessfully");
+        Swal.fire("Added Sucessfully");
       this.loadModal();
       this.viewData();
       }
@@ -80,12 +80,32 @@ export class InterestedComponent implements OnInit {
 
   // To delete bloodgroup
   deleteBloodgroup(id: any) {
-    this.request.deleteInterests(id).subscribe(res => {
-      console.log(id);
-      this.viewData();
-    console.log('Deleted');
-    this.router.navigate(['interest']);
-    });
+
+    Swal.fire({  
+      title: 'Are you sure want to Delete?',  
+      text: 'You will not be able to recover this Data',  
+      icon: 'warning',  
+      showCancelButton: true,  
+      confirmButtonText: 'Delete',  
+      cancelButtonText: 'Cancel'  
+    }).then((result) => {  
+      if (result.value) {   
+        this.request.deleteInterests(id).subscribe(res => {
+          console.log(id);
+          this.viewData();
+        console.log('Deleted');
+        });
+        Swal.fire(  
+          'Deleted! Sucessfully',  
+        )  
+      } else if (result.dismiss === Swal.DismissReason.cancel) {  
+        Swal.fire(  
+          'Cancelled',   
+        )  
+      }  
+    }) 
+
+  
   }
 
   // To edit bloodgroup
@@ -115,7 +135,7 @@ export class InterestedComponent implements OnInit {
     }
   this.request.updateInterests(this.IdValue,edata).subscribe((res : any) => {
     if (res.status == 'success') {
-      swal("Updated Sucessfully");
+      Swal.fire("Updated Sucessfully");
       this.loadModal();
       this.viewData();
     }
